rely on git symbolic-ref to shorten ref
Prior to this commit Git-RSL would get itself into a weird state if the
current branch name contained a `/`, which is a legal character in a
name, because the `sed` would take only the section after the last `/`
as the branch name even though 1) it might not be a valid ref and 2) it
might be a _different_ ref. This commit changes both git-securefetch and
git-securepush to rely on the `--short` flag of `git symbolic-ref`,
which has existed since at least git 1.9.4, to do the ref name
shortening.